added exceptions

+++ BackEnd/src/main/java/com/example/Validation/controllers/UserController.java
@@ -14,6 +14,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.Validation.exceptions.DuplicateEmailException;
+import com.example.Validation.exceptions.DuplicateUsernameException;
+import com.example.Validation.exceptions.EmailValidationException;
+import com.example.Validation.exceptions.UserNotFoundException;
 import com.example.Validation.models.UserModel;
 import com.example.Validation.repositories.UserRepository;
 
@@ -29,9 +33,13 @@ public class UserController {
 	}
 	@CrossOrigin
 	@PostMapping("/register")
-	public void registerUser(@RequestBody UserModel data){
+	public void registerUser(@RequestBody UserModel data) throws EmailValidationException, DuplicateEmailException, DuplicateUsernameException{
 		if(!validateEmail(data.getEmail()))
-			return;
+			throw new EmailValidationException();
+		if(!userRepository.findByEmail(data.getEmail()).isEmpty())
+			throw new DuplicateEmailException();
+		if(!userRepository.findByUsername(data.getUsername()).isEmpty())
+			throw new DuplicateUsernameException();
 		userRepository.save_user(data);
 	}
 	
@@ -43,27 +51,41 @@ public class UserController {
 	}
 	@CrossOrigin
 	@GetMapping("/getuserbyemail")
-	public ResponseEntity<?> getUserByEmail(@RequestParam String email) {
+	public ResponseEntity<?> getUserByEmail(@RequestParam String email) throws UserNotFoundException {
 		List<UserModel> users=userRepository.findByEmail(email);
-		System.out.println(users.get(0).getUsername());
+		if(users.isEmpty()) {
+			throw new UserNotFoundException();
+		}
 		return new ResponseEntity<>(users.get(0),HttpStatus.OK);
 	}
+	
 	@CrossOrigin
 	@GetMapping("/getRole")
-	public ResponseEntity<?> getRole(@RequestParam String email) {
+	public ResponseEntity<?> getRole(@RequestParam String email) throws UserNotFoundException {
 		List<UserModel> users=userRepository.findByEmail(email);
+		if(users.isEmpty()) {
+			throw new UserNotFoundException();
+		}
 		return new ResponseEntity<>(users.get(0).getRole(),HttpStatus.OK);
 	}
 	@CrossOrigin
 	@GetMapping("/activate")
-	public ResponseEntity<?> activateAccount(@RequestParam String username) {
-		userRepository.activateAccount(username);
+	public ResponseEntity<?> activateAccount(@RequestParam String email) throws UserNotFoundException {
+		List<UserModel> users=userRepository.findByEmail(email);
+		if(users.isEmpty()) {
+			throw new UserNotFoundException();
+		}
+		userRepository.activateAccount(email);
 		return new ResponseEntity<>(true,HttpStatus.OK);
 	}
 	
 	@CrossOrigin
 	@GetMapping("/delete")
-	public ResponseEntity<?> deleteAccount(@RequestParam String email) {
+	public ResponseEntity<?> deleteAccount(@RequestParam String email) throws UserNotFoundException {
+		List<UserModel> users=userRepository.findByEmail(email);
+		if(users.isEmpty()) {
+			throw new UserNotFoundException();
+		}
 		userRepository.deleteAccount(email);
 		return new ResponseEntity<>(true,HttpStatus.OK);
 	}
@@ -85,15 +107,6 @@ public class UserController {
 	public ResponseEntity<?> updateUser(@RequestBody UserModel data) {
 		userRepository.updateAccount(data);
 		return new ResponseEntity<>(true,HttpStatus.OK);
-	}
-	@CrossOrigin
-	@PostMapping("/finduser")
-	public ResponseEntity<?> findUsers(@RequestBody UserModel data) {
-		boolean found=false;
-		if(userRepository.findByEmail(data.getEmail()).size()!=0) {
-			found=true;
-		} 
-		return new ResponseEntity<>(found,HttpStatus.CREATED);
 	}
